% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/council_plot_ly.R
\name{council_plot_ly}
\alias{council_plot_ly}
\title{Plot_ly object, with custom styling}
\usage{
council_plot_ly(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[plotly:plot_ly]{plotly::plot_ly()}}}
}
\value{
\code{\link[plotly:plot_ly]{plotly::plot_ly()}} object
}
\description{
Plot_ly object, with custom styling
}
\examples{
\dontrun{
council_plot_ly()
}
}
