% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datatable_container.R
\name{create_datatable_container}
\alias{create_datatable_container}
\title{Create HTML container for Two-way table}
\usage{
create_datatable_container(
  twoway_data,
  type = c("proportion", "proportion_w_se", "total", "total_w_se", "sample")
)
}
\arguments{
\item{twoway_data}{list, return from \code{create_twoway_table()}}

\item{type}{character, one of \code{c("proportion", "proportion_w_se","total", "total_w_se", "sample")}}
}
\value{
named list, of two items.
- \code{dt_data} table data
- \code{container} HTML DT container
}
\description{
Create HTML container for Two-way table
}
