% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_one_way_table.R
\name{create_one_way_table}
\alias{create_one_way_table}
\title{Create a one-way cross table}
\usage{
create_one_way_table(variable_row, hh_ids)
}
\arguments{
\item{variable_row}{character, variable name. Must be one of \code{tbi_dict$variable}.}

\item{hh_ids}{vector, integer64 vector of unique household IDs. Must be
one or more of \code{tbi_tables$hh$hhid}}
}
\value{
A named list of tables
- \code{table}, table with columns
- variable  name, matching \code{variable_row}
- \code{group_N} raw sample size - number of people, trips, households, days (by group)
- \code{group_N_hh} number of households in sample (by group)
- \code{expanded_total} and \code{expanded_total_se}
- \code{estimated_prop} and \code{estimated_prop_se} estimated proportion (0.0 - 1.0) and SE;
multiply by 100 for percentages.
- \code{definitions} table with contextual information from \code{tbi_dict}
- \code{summary}, table
}
\description{
Create a one-way cross table
}
\examples{
\dontrun{

library(travel.survey.explorer)
create_one_way_table("bike_freq")
}
}
