% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_joining_moe.R
\name{create_table_joining_moe}
\alias{create_table_joining_moe}
\title{Dimension related tables}
\usage{
create_table_joining_moe(valuetable, moetable, xalias, xvalues)
}
\arguments{
\item{valuetable}{value table}

\item{moetable}{margin of error table}

\item{xalias}{x alias}

\item{xvalues}{x values}
}
\value{
a \code{\link[data.table:data.table]{data.table::data.table()}}
}
\description{
Dimension related tables
}
