% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_vistable_moe.R
\name{create_table_vistable_moe}
\alias{create_table_vistable_moe}
\title{Title}
\usage{
create_table_vistable_moe(valuetable, moetable, xalias, xvalues)
}
\arguments{
\item{valuetable}{value table}

\item{moetable}{Margin of error table}

\item{xalias}{x var alias}

\item{xvalues}{x var values}
}
\value{
a \code{\link[data.table:data.table]{data.table::data.table()}}
}
\description{
Title
}
