% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_container_dtstyle.R
\name{dt_container_dtstyle}
\alias{dt_container_dtstyle}
\title{Style DT}
\usage{
dt_container_dtstyle(atable, xvaralias, yvaralias)
}
\arguments{
\item{atable}{a table}

\item{xvaralias}{character, x var alias}

\item{yvaralias}{character, y var alias}
}
\value{
HTML styling
}
\description{
Style DT
}
\seealso{
Other DT styling: 
\code{\link{dt_container_tblMOE_dtstyle}()}
}
\concept{DT styling}
