% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_container_tblMOE_dtstyle.R
\name{dt_container_tblMOE_dtstyle}
\alias{dt_container_tblMOE_dtstyle}
\title{Style DT with MOE}
\usage{
dt_container_tblMOE_dtstyle(
  atable,
  xvaralias,
  yvaralias,
  tbltype = c("share", "estimate", "mean")
)
}
\arguments{
\item{atable}{a table}

\item{xvaralias}{x var alias}

\item{yvaralias}{y var alias}

\item{tbltype}{character, one of \code{"share"}, \code{"estimate"}, or \code{"mean"}}
}
\value{
HTML styling
}
\description{
Style DT with MOE
}
\seealso{
Other DT styling: 
\code{\link{dt_container_dtstyle}()}
}
\concept{DT styling}
