% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_list}
\alias{input_list}
\title{Input list}
\format{
named list of possible user inputs
}
\usage{
input_list
}
\description{
Input list
}
\seealso{
Other data: 
\code{\link{input_question_list}},
\code{\link{missing_codes}},
\code{\link{pii_codes}},
\code{\link{plotly_layout}},
\code{\link{tbi_dict}},
\code{\link{tbi_tables}}
}
\concept{data}
\keyword{datasets}
