% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_question_list}
\alias{input_question_list}
\title{Input question list}
\format{
named list for fetching the corresponding survey question for a given
variable.
}
\usage{
input_question_list
}
\description{
Input question list
}
\seealso{
Other data: 
\code{\link{input_list}},
\code{\link{missing_codes}},
\code{\link{pii_codes}},
\code{\link{plotly_layout}},
\code{\link{tbi_dict}},
\code{\link{tbi_tables}}
}
\concept{data}
\keyword{datasets}
