% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_codes}
\alias{missing_codes}
\title{Missing data codes}
\format{
vector, 10 character values indicating that data is missing
}
\usage{
missing_codes
}
\description{
Missing data codes
}
\seealso{
Other data: 
\code{\link{input_list}},
\code{\link{input_question_list}},
\code{\link{pii_codes}},
\code{\link{plotly_layout}},
\code{\link{tbi_dict}},
\code{\link{tbi_tables}}
}
\concept{data}
\keyword{datasets}
