% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input_category_variable.R
\name{mod_input_category_variable_ui}
\alias{mod_input_category_variable_ui}
\title{input_category_variable UI Function}
\usage{
mod_input_category_variable_ui(
  id,
  starting_category_selected = "Demographics",
  starting_variable_choices = input_list$Demographics,
  starting_variable_selected = "Age"
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{starting_category_selected}{character, starting category}

\item{starting_variable_choices}{list, starting variable choices. Default is
\code{input_list$Demographics}}

\item{starting_variable_selected}{character, starting variable selection.
Default is \code{"Age"}}
}
\description{
A shiny Module.
}
\note{
Category and variable must be corresponding.

\if{html}{\out{<div class="sourceCode">}}\preformatted{
# good
mod_input_category_variable_ui(id = "an_id",
    starting_category_selected = "Demographics",
    starting_variable_choices = input_list$Demographics,
    starting_variable_selected = "Age")

# bad
mod_input_category_variable_ui(id = "an_id",
    starting_category_selected = "Demographics",
    starting_variable_choices = input_list$Trips,
    starting_variable_selected = "City of Trip Origin")

}\if{html}{\out{</div>}}
}
