% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pii_codes}
\alias{pii_codes}
\title{Personally identifiable information}
\format{
vector, 26 character values indicating sensitive data. These data are removed from \code{tbi_tables}
}
\usage{
pii_codes
}
\description{
Personally identifiable information
}
\seealso{
Other data: 
\code{\link{input_list}},
\code{\link{input_question_list}},
\code{\link{missing_codes}},
\code{\link{plotly_layout}},
\code{\link{tbi_dict}},
\code{\link{tbi_tables}}
}
\concept{data}
\keyword{datasets}
