% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plotly_layout}
\alias{plotly_layout}
\title{Plotly formatting}
\format{
nested, named list with specifications for \code{\link[plotly:layout]{plotly::layout()}}
}
\usage{
plotly_layout
}
\description{
Plotly formatting
}
\examples{
library(travel.survey.explorer)
plotly_layout$hover_text$size

}
\seealso{
Other data: 
\code{\link{input_list}},
\code{\link{input_question_list}},
\code{\link{missing_codes}},
\code{\link{pii_codes}},
\code{\link{tbi_dict}},
\code{\link{tbi_tables}}
}
\concept{data}
\keyword{datasets}
