% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbi_dict}
\alias{tbi_dict}
\title{Travel Behavior Inventory Dictionary}
\format{
data.frame with 10 columns and 904 rows. Contains contextual
and meta data o
- \code{which_table} character,
one of \code{c("per", "trip", "day", "veh", "hh", "trip_purpose")}
- \code{category} character,
one of  \code{c("Demographics", "Attitudes toward autonomous vehicles", "Shared mobility",  "Commute", "Trips", "Days without travel", "Delivery & online shopping", "Vehicle")}
- \code{variable} character, short form variable code in snake case
- \code{survey_question} character, survey question as it was asked to the respondent
- \code{variable_logic} character, any associated logic. Some questions were only asked
if a respondent had a certain answer to a previous question.
- \verb{variable label} character, person-readable variable label.
- \code{value} numeric, potential response value
- \code{value_label} character, label corresponding to response value
- \code{value_logic} character, any associated logic. Some questions were only asked
if a respondent had a certain answer to a previous question.
- \code{wt_field} character, one of \code{c("person_weight", "trip_weight", "day_weight", "hh_weight", "trip_purpose_weight")}
}
\usage{
tbi_dict
}
\description{
Travel Behavior Inventory Dictionary
}
\seealso{
Other data: 
\code{\link{input_list}},
\code{\link{input_question_list}},
\code{\link{missing_codes}},
\code{\link{pii_codes}},
\code{\link{plotly_layout}},
\code{\link{tbi_tables}}
}
\concept{data}
\keyword{datasets}
