% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbi_tables}
\alias{tbi_tables}
\title{Travel Behavior Inventory data tables}
\format{
named list of six tables for each type of TBI data.
- \code{day} - Day table
- \code{per}
- \code{hh}
- \code{veh}'
- \code{trip}
- \code{trip_purpose}
}
\usage{
tbi_tables
}
\description{
Travel Behavior Inventory data tables
}
\seealso{
Other data: 
\code{\link{input_list}},
\code{\link{input_question_list}},
\code{\link{missing_codes}},
\code{\link{plotly_layout}},
\code{\link{tbi_dict}}
}
\concept{data}
\keyword{datasets}
