% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.survey.explorer-package.R
\docType{package}
\name{travel.survey.explorer-package}
\alias{travel.survey.explorer}
\alias{travel.survey.explorer-package}
\title{travel.survey.explorer: Travel Survey Explorer}
\description{
Interactive data visualization for the Metropolitan Council Travel Behavior Inventory
}
\author{
\strong{Maintainer}: Liz Roten \email{liz.roten@metc.state.mn.us} (\href{https://orcid.org/0000-0002-5346-3549}{ORCID})

Authors:
\itemize{
  \item Ashley Asmus \email{ashley.asmus@metc.state.mn.us} (\href{https://orcid.org/0000-0001-5505-1372}{ORCID})
}

Other contributors:
\itemize{
  \item Metropolitan Council [copyright holder]
}

}
\keyword{internal}
