% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Select variables with a function}
\usage{
where(fn)
}
\arguments{
\item{fn}{A function that returns \code{TRUE} or \code{FALSE} (technically, a
\emph{predicate} function). Can also be a purrr-like formula.}
}
\description{
This code is copied directly from https://github.com/r-lib/tidyselect/blob/HEAD/R/helpers-where.R

This \link[=language]{selection helper} selects the variables for which a
function returns \code{TRUE}.
}
