% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_col_subset.R
\name{xtab_col_subset}
\alias{xtab_col_subset}
\title{Subset cross tab columns}
\usage{
xtab_col_subset(table, colstring = col.headers)
}
\arguments{
\item{table}{table}

\item{colstring}{defautl is col.headers}
}
\value{
a \code{\link[data.table:data.table]{data.table::data.table()}}
}
\description{
Subset cross tab columns
}
\seealso{
Other shiny processing: 
\code{\link{xtab_join_samplecnt}()}
}
\concept{shiny processing}
