% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_create_DT.R
\name{xtab_create_DT}
\alias{xtab_create_DT}
\title{Create DT}
\usage{
xtab_create_DT(
  atable,
  moe = c(TRUE, FALSE),
  acontainer,
  indices2hide,
  maxyvals,
  sc.cols
)
}
\arguments{
\item{atable}{a table}

\item{moe}{logical, included margin of error}

\item{acontainer}{container}

\item{indices2hide}{columns to hide}

\item{maxyvals}{numeric, maximum number of columns}

\item{sc.cols}{numeric, columns with actual data}
}
\value{
a \code{\link[DT:datatable]{DT::datatable()}} object
}
\description{
Create DT
}
