% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_join_samplecnt.R
\name{xtab_join_samplecnt}
\alias{xtab_join_samplecnt}
\title{Join crosstab with relevant sample count}
\usage{
xtab_join_samplecnt(xtabcleandt, dttype, varsXAlias)
}
\arguments{
\item{xtabcleandt}{crosstab, clean}

\item{dttype}{data table type}

\item{varsXAlias}{X variable alias}
}
\value{
a \code{\link[data.table:data.table]{data.table::data.table()}}
}
\description{
Join crosstab with relevant sample count
}
\seealso{
Other shiny processing: 
\code{\link{xtab_col_subset}()}
}
\concept{shiny processing}
