% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_tblMOE_join_samplecnt.R
\name{xtab_tblMOE_join_samplecnt}
\alias{xtab_tblMOE_join_samplecnt}
\title{Join crosstab, MOE, sample count data}
\usage{
xtab_tblMOE_join_samplecnt(xtabcleantblMOEdt, xtabcleandt, dttype, varsXAlias)
}
\arguments{
\item{xtabcleantblMOEdt}{crosstab, clean, with margin of error}

\item{xtabcleandt}{crosstab clean}

\item{dttype}{crosstab data type}

\item{varsXAlias}{X variable alias}
}
\value{
a \code{\link[data.table:data.table]{data.table::data.table()}}
}
\description{
Join crosstab, MOE, sample count data
}
