% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_one_way_table.R
\name{create_one_way_table}
\alias{create_one_way_table}
\title{Create a one-way cross table}
\usage{
create_one_way_table(variable)
}
\arguments{
\item{variable}{character, variable name.}

\item{type}{character, one of \code{"numeric"} or \code{"category"}.
Default is \code{"numeric"}}
}
\value{
A tibble with columns
\itemize{
\item variable name with relevant values
\item \code{group_N} raw sample size - number of people, trips, households, days (by group)
\item \code{group_N_hh} number of households in sample (by group)
\item \code{expanded_total} and \code{expanded_total_se}
\item \code{estimated_prop} and \code{estimated_prop_se} estimated proportion (0.0 - 1.0) and SE;
multiply by 100 for percentages.
}
}
\description{
Create a one-way cross table
}
\examples{
/dontrun{

library(travel.survey.explorer)
create_one_way_table("bike_freq")

}
}
